Implemented virtual path for media module.
First test implementation done in the Helper module.
The upload component is using a bad workaround where the virtual path is defined
in the uri as ?virutalPath= because we are not uploading media files AND form data
at the same time. Therefore the virtualPath can currently only be passed in the get/uri data.